refactor: split client code from server code
While implementing support for ModMenu I overlooked the fact that the `SleepWarp` class implements the `ModInitializer` interface, which causes a crash on the dedicated server since YACL and ModMenu code are implemented within that class.
To remedy this I decided to create two nested classes which implement `ClientModInitializer` and `DedicatedServerModInitializer ` respectively; `SleepWarp.Client` and `SleepWarp.Server`. Each class will handle logic for its side of the game.

Additionally, I iterated further on `SleepWarp.Config` in this commit.