refactor: swap method refs to lambda for config getters
Using method references will cause stale values to be retrieved by YACL when attempting to save to disk which results in unusual behavior. I'm unsure if this is an issue caused by YACL or a weird Java quirk; the YACL developer didn't seem to care much about the issue, so I suppose we'll never know.