refactor: split config and tick monitor from main class
The SleepWarp class was getting a bit confusing while reading and editing, so I've split things up a bit.
- Moved SleepWarp.Config into its own JsonConfiguration class
- Moved SleepWarp.TickMonitor into its own TickMonitor class
- Added static variables TICK_MONITOR and CONFIGURATION to SleepWarp, each with an instance of TickMonitor and JsonConfiguration respectively

Additionally, several changes were made to ServerWorldMixin to wake players on the exact tick of the new day and an attempt was made to improve performance-based tick scaling. These changes resulted in the removal of SleepManagerAccessor as it was no longer needed.