bug: fix block duplication when tick_random_block is true
My port for 1.20 was made in haste and I botched the random block tick implemenation which introduced a major bug which caused any block which could be random ticked to be duplicated at 16 block intervals through the depth of the chunk it was in.

I've located my error and corrected it, and cleaned a few things up.