refactor: improve warp performance
My multithreading implementation in WarpDrive left a lot to be desired. Although it was an improvement from my last attempt this implementation had high memory usage along with low throughput which caused to a large amount of warp ticks to be executed well after the day had arrived and the player is awake resulting in crops rapidly growing before the players eyes.

To try to improve this, I've ditched WarpDrive and migrated its core functionality to the new WarpEngine class and all of the world tick logic into dedicated Runnable classes.
Instead of having three threads dedicated to different aspects of world ticking, each part of the world tick will be executed asynchronously on an available thread from the ForkJoinPool. This new implementation has a huge increase in execution throughput which in turn causes the majority of world warp ticks to complete a few world ticks later.

I also added a check to make sure that worlds which don't allow for sleeping don't get their time accelerated, and tweaked the action bar text.